Adds logic to place order and control Max Available of products based on orders placed.

+++ shoppe-api/Shoppe.Api.UnitTest/Validators/ProductValidatorTests.cs
@@ -12,7 +12,7 @@ namespace Shoppe.Api.UnitTest.Validators
         [InlineData("Milk", "Milk Desc", "image path", -3, 3)]
         [InlineData("Milk", "Milk Desc", "image path", 5, -3)]
         [InlineData("Milk", "Milk Desc", "image path", 5, 0, true)]
-        public void ProductValidator_Fails_On_Invalid_Input(string code, string description, string imageUrl, float price, float qty, bool shouldPass = false)
+        public void ProductValidator_Fails_On_Invalid_Input(string code, string description, string imageUrl, float price, int qty, bool shouldPass = false)
         {
             // Arrange
             var product = new Product(code, description, imageUrl, price, qty);
